requirements.txt
# RA-YOLO: 遥感飞机旋转目标检测项目依赖
# Remote Sensing Aircraft Detection with Oriented Bounding Boxes

ultralytics>=8.0.0
torch>=1.8.0
torchvision>=0.9.0
opencv-python-headless>=4.5.0
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
albumentations>=1.0.0
Pillow>=8.0.0
scipy>=1.7.0
PyYAML>=5.4.0
tqdm>=4.60.0
seaborn>=0.11.0
scikit-learn>=0.24.0
